requirements0.txt
streamlit==1.12.0
pandas==1.4.3
python-dotenv==0.20.0
